+++ packages/compiler-cli/ngcc/src/writing/file_writer.ts
@@ -8,12 +8,13 @@
  */
 import {EntryPoint} from '../packages/entry_point';
 import {EntryPointBundle} from '../packages/entry_point_bundle';
-import {FileInfo} from '../rendering/renderer';
+import {FileToWrite} from '../rendering/utils';
 
 
 /**
  * Responsible for writing out the transformed files to disk.
  */
 export interface FileWriter {
-  writeBundle(entryPoint: EntryPoint, bundle: EntryPointBundle, transformedFiles: FileInfo[]): void;
+  writeBundle(entryPoint: EntryPoint, bundle: EntryPointBundle, transformedFiles: FileToWrite[]):
+      void;
 }
