+++ packages/compiler-cli/ngcc/test/rendering/esm5_rendering_formatter_spec.ts
@@ -14,9 +14,9 @@ import {DecorationAnalyzer} from '../../src/analysis/decoration_analyzer';
 import {NgccReferencesRegistry} from '../../src/analysis/ngcc_references_registry';
 import {SwitchMarkerAnalyzer} from '../../src/analysis/switch_marker_analyzer';
 import {IMPORT_PREFIX} from '../../src/constants';
-import {Esm2015ReflectionHost} from '../../src/host/esm2015_host';
-import {EsmRenderer} from '../../src/rendering/esm_renderer';
-import {makeTestEntryPointBundle} from '../helpers/utils';
+import {Esm5ReflectionHost} from '../../src/host/esm5_host';
+import {Esm5RenderingFormatter} from '../../src/rendering/esm5_rendering_formatter';
+import {makeTestEntryPointBundle, getDeclaration} from '../helpers/utils';
 import {MockFileSystem} from '../helpers/mock_file_system';
 import {MockLogger} from '../helpers/mock_logger';
 
@@ -25,21 +25,22 @@ const _ = AbsoluteFsPath.fromUnchecked;
 function setup(file: {name: AbsoluteFsPath, contents: string}) {
   const fs = new MockFileSystem();
   const logger = new MockLogger();
-  const bundle = makeTestEntryPointBundle('es2015', 'esm2015', false, [file]) !;
+  const bundle = makeTestEntryPointBundle('module', 'esm5', false, [file]);
   const typeChecker = bundle.src.program.getTypeChecker();
-  const host = new Esm2015ReflectionHost(logger, false, typeChecker);
+  const host = new Esm5ReflectionHost(logger, false, typeChecker);
   const referencesRegistry = new NgccReferencesRegistry(host);
-  const decorationAnalyses = new DecorationAnalyzer(
-                                 fs, bundle.src.program, bundle.src.options, bundle.src.host,
-                                 typeChecker, host, referencesRegistry, [_('/')], false)
-                                 .analyzeProgram();
+  const decorationAnalyses =
+      new DecorationAnalyzer(
+          fs, bundle.src.program, bundle.src.options, bundle.src.host, typeChecker, host,
+          referencesRegistry, [AbsoluteFsPath.fromUnchecked('/')], false)
+          .analyzeProgram();
   const switchMarkerAnalyses = new SwitchMarkerAnalyzer(host).analyzeProgram(bundle.src.program);
-  const renderer = new EsmRenderer(fs, logger, host, false, bundle);
+  const renderer = new Esm5RenderingFormatter(host, false);
   const importManager = new ImportManager(new NoopImportRewriter(), IMPORT_PREFIX);
   return {
     host,
     program: bundle.src.program,
-    sourceFile: bundle.src.file, renderer, decorationAnalyses, switchMarkerAnalyses, importManager,
+    sourceFile: bundle.src.file, renderer, decorationAnalyses, switchMarkerAnalyses, importManager
   };
 }
 
@@ -49,26 +50,49 @@ const PROGRAM = {
 /* A copyright notice */
 import 'some-side-effect';
 import {Directive} from '@angular/core';
-export class A {}
-A.decorators = [
-  { type: Directive, args: [{ selector: '[a]' }] },
-  { type: OtherA }
-];
-export class B {}
-B.decorators = [
-  { type: OtherB },
-  { type: Directive, args: [{ selector: '[b]' }] }
-];
-export class C {}
-C.decorators = [
-  { type: Directive, args: [{ selector: '[c]' }] },
-];
-let compileNgModuleFactory = compileNgModuleFactory__PRE_R3__;
-let badlyFormattedVariable = __PRE_R3__badlyFormattedVariable;
+var A = (function() {
+  function A() {}
+  A.decorators = [
+    { type: Directive, args: [{ selector: '[a]' }] },
+    { type: OtherA }
+  ];
+  A.prototype.ngDoCheck = function() {
+    //
+  };
+  return A;
+}());
 
+var B = (function() {
+  function B() {}
+  B.decorators = [
+    { type: OtherB },
+    { type: Directive, args: [{ selector: '[b]' }] }
+  ];
+  return B;
+}());
+
+var C = (function() {
+  function C() {}
+  C.decorators = [
+    { type: Directive, args: [{ selector: '[c]' }] },
+  ];
+  return C;
+}());
+
+function NoIife() {}
+
+var BadIife = (function() {
+  function BadIife() {}
+  BadIife.decorators = [
+    { type: Directive, args: [{ selector: '[c]' }] },
+  ];
+}());
+
+var compileNgModuleFactory = compileNgModuleFactory__PRE_R3__;
+var badlyFormattedVariable = __PRE_R3__badlyFormattedVariable;
 function compileNgModuleFactory__PRE_R3__(injector, options, moduleType) {
-  const compilerFactory = injector.get(CompilerFactory);
-  const compiler = compilerFactory.createCompiler([options]);
+  var compilerFactory = injector.get(CompilerFactory);
+  var compiler = compilerFactory.createCompiler([options]);
   return compiler.compileModuleAsync(moduleType);
 }
 
@@ -76,48 +100,62 @@ function compileNgModuleFactory__POST_R3__(injector, options, moduleType) {
   ngDevMode && assertNgModuleType(moduleType);
   return Promise.resolve(new R3NgModuleFactory(moduleType));
 }
-// Some other content`
+// Some other content
+export {A, B, C, NoIife, BadIife};`
 };
 
 const PROGRAM_DECORATE_HELPER = {
   name: _('/some/file.js'),
   contents: `
 import * as tslib_1 from "tslib";
-var D_1;
 /* A copyright notice */
 import { Directive } from '@angular/core';
-const OtherA = () => (node) => { };
-const OtherB = () => (node) => { };
-let A = class A {
-};
-A = tslib_1.__decorate([
-    Directive({ selector: '[a]' }),
-    OtherA()
-], A);
+var OtherA = function () { return function (node) { }; };
+var OtherB = function () { return function (node) { }; };
+var A = /** @class */ (function () {
+    function A() {
+    }
+    A = tslib_1.__decorate([
+        Directive({ selector: '[a]' }),
+        OtherA()
+    ], A);
+    return A;
+}());
 export { A };
-let B = class B {
-};
-B = tslib_1.__decorate([
-    OtherB(),
-    Directive({ selector: '[b]' })
-], B);
+var B = /** @class */ (function () {
+    function B() {
+    }
+    B = tslib_1.__decorate([
+        OtherB(),
+        Directive({ selector: '[b]' })
+    ], B);
+    return B;
+}());
 export { B };
-let C = class C {
-};
-C = tslib_1.__decorate([
-    Directive({ selector: '[c]' })
-], C);
+var C = /** @class */ (function () {
+    function C() {
+    }
+    C = tslib_1.__decorate([
+        Directive({ selector: '[c]' })
+    ], C);
+    return C;
+}());
 export { C };
-let D = D_1 = class D {
-};
-D = D_1 = tslib_1.__decorate([
-    Directive({ selector: '[d]', providers: [D_1] })
-], D);
+var D = /** @class */ (function () {
+    function D() {
+    }
+    D_1 = D;
+    var D_1;
+    D = D_1 = tslib_1.__decorate([
+        Directive({ selector: '[d]', providers: [D_1] })
+    ], D);
+    return D;
+}());
 export { D };
 // Some other content`
 };
 
-describe('Esm2015Renderer', () => {
+describe('Esm5RenderingFormatter', () => {
 
   describe('addImports', () => {
     it('should insert the given imports after existing imports of the source file', () => {
@@ -152,7 +190,7 @@ import * as i1 from '@angular/common';`);
           ],
           importManager, sourceFile);
       expect(output.toString()).toContain(`
-// Some other content
+export {A, B, C, NoIife, BadIife};
 export {ComponentA1} from './a';
 export {ComponentA2} from './a';
 export {ComponentB} from './foo/b';
@@ -165,9 +203,9 @@ export {TopLevelComponent};`);
       renderer.addExports(
           output, _(PROGRAM.name.replace(/\.js$/, '')),
           [
-            {from: _('/some/a.js'), alias: 'eComponentA1', identifier: 'ComponentA1'},
-            {from: _('/some/a.js'), alias: 'eComponentA2', identifier: 'ComponentA2'},
-            {from: _('/some/foo/b.js'), alias: 'eComponentB', identifier: 'ComponentB'},
+            {from: _('/some/a.js'), alias: _('eComponentA1'), identifier: 'ComponentA1'},
+            {from: _('/some/a.js'), alias: _('eComponentA2'), identifier: 'ComponentA2'},
+            {from: _('/some/foo/b.js'), alias: _('eComponentB'), identifier: 'ComponentB'},
             {from: PROGRAM.name, alias: 'eTopLevelComponent', identifier: 'TopLevelComponent'},
           ],
           importManager, sourceFile);
@@ -186,12 +224,12 @@ export {TopLevelComponent};`);
         throw new Error(`Could not find source file`);
       }
       const output = new MagicString(PROGRAM.contents);
-      renderer.addConstants(output, 'const x = 3;', file);
+      renderer.addConstants(output, 'var x = 3;', file);
       expect(output.toString()).toContain(`
 import {Directive} from '@angular/core';
 
-const x = 3;
-export class A {}`);
+var x = 3;
+var A = (function() {`);
     });
 
     it('should insert constants after inserted imports', () => {
@@ -201,20 +239,20 @@ export class A {}`);
         throw new Error(`Could not find source file`);
       }
       const output = new MagicString(PROGRAM.contents);
-      renderer.addConstants(output, 'const x = 3;', file);
+      renderer.addConstants(output, 'var x = 3;', file);
       renderer.addImports(output, [{specifier: '@angular/core', qualifier: 'i0'}], file);
       expect(output.toString()).toContain(`
 import {Directive} from '@angular/core';
 import * as i0 from '@angular/core';
 
-const x = 3;
-export class A {`);
+var x = 3;
+var A = (function() {`);
     });
   });
 
   describe('rewriteSwitchableDeclarations', () => {
     it('should switch marked declaration initializers', () => {
-      const {renderer, program, switchMarkerAnalyses, sourceFile} = setup(PROGRAM);
+      const {renderer, program, sourceFile, switchMarkerAnalyses} = setup(PROGRAM);
       const file = program.getSourceFile('some/file.js');
       if (file === undefined) {
         throw new Error(`Could not find source file`);
@@ -223,11 +261,11 @@ export class A {`);
       renderer.rewriteSwitchableDeclarations(
           output, file, switchMarkerAnalyses.get(sourceFile) !.declarations);
       expect(output.toString())
-          .not.toContain(`let compileNgModuleFactory = compileNgModuleFactory__PRE_R3__;`);
+          .not.toContain(`var compileNgModuleFactory = compileNgModuleFactory__PRE_R3__;`);
       expect(output.toString())
-          .toContain(`let badlyFormattedVariable = __PRE_R3__badlyFormattedVariable;`);
+          .toContain(`var badlyFormattedVariable = __PRE_R3__badlyFormattedVariable;`);
       expect(output.toString())
-          .toContain(`let compileNgModuleFactory = compileNgModuleFactory__POST_R3__;`);
+          .toContain(`var compileNgModuleFactory = compileNgModuleFactory__POST_R3__;`);
       expect(output.toString())
           .toContain(`function compileNgModuleFactory__PRE_R3__(injector, options, moduleType) {`);
       expect(output.toString())
@@ -236,81 +274,100 @@ export class A {`);
   });
 
   describe('addDefinitions', () => {
-    it('should insert the definitions directly after the class declaration', () => {
-      const {renderer, decorationAnalyses, sourceFile} = setup(PROGRAM);
-      const output = new MagicString(PROGRAM.contents);
-      const compiledClass =
-          decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'A') !;
-      renderer.addDefinitions(output, compiledClass, 'SOME DEFINITION TEXT');
-      expect(output.toString()).toContain(`
-export class A {}
+    it('should insert the definitions directly before the return statement of the class IIFE',
+       () => {
+         const {renderer, decorationAnalyses, sourceFile} = setup(PROGRAM);
+         const output = new MagicString(PROGRAM.contents);
+         const compiledClass =
+             decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'A') !;
+         renderer.addDefinitions(output, compiledClass, 'SOME DEFINITION TEXT');
+         expect(output.toString()).toContain(`
+  A.prototype.ngDoCheck = function() {
+    //
+  };
 SOME DEFINITION TEXT
-A.decorators = [
+  return A;
 `);
-    });
+       });
 
+    it('should error if the compiledClass is not valid', () => {
+      const {renderer, sourceFile, program} = setup(PROGRAM);
+      const output = new MagicString(PROGRAM.contents);
+
+      const noIifeDeclaration =
+          getDeclaration(program, sourceFile.fileName, 'NoIife', ts.isFunctionDeclaration);
+      const mockNoIifeClass: any = {declaration: noIifeDeclaration, name: _('NoIife')};
+      expect(() => renderer.addDefinitions(output, mockNoIifeClass, 'SOME DEFINITION TEXT'))
+          .toThrowError(
+              'Compiled class declaration is not inside an IIFE: NoIife in /some/file.js');
+
+      const badIifeDeclaration =
+          getDeclaration(program, sourceFile.fileName, 'BadIife', ts.isVariableDeclaration);
+      const mockBadIifeClass: any = {declaration: badIifeDeclaration, name: _('BadIife')};
+      expect(() => renderer.addDefinitions(output, mockBadIifeClass, 'SOME DEFINITION TEXT'))
+          .toThrowError(
+              'Compiled class wrapper IIFE does not have a return statement: BadIife in /some/file.js');
+    });
   });
 
 
   describe('removeDecorators', () => {
-    describe('[static property declaration]', () => {
-      it('should delete the decorator (and following comma) that was matched in the analysis',
-         () => {
-           const {decorationAnalyses, sourceFile, renderer} = setup(PROGRAM);
-           const output = new MagicString(PROGRAM.contents);
-           const compiledClass =
-               decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'A') !;
-           const decorator = compiledClass.decorators[0];
-           const decoratorsToRemove = new Map<ts.Node, ts.Node[]>();
-           decoratorsToRemove.set(decorator.node.parent !, [decorator.node]);
-           renderer.removeDecorators(output, decoratorsToRemove);
-           expect(output.toString())
-               .not.toContain(`{ type: Directive, args: [{ selector: '[a]' }] },`);
-           expect(output.toString()).toContain(`{ type: OtherA }`);
-           expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[b]' }] }`);
-           expect(output.toString()).toContain(`{ type: OtherB }`);
-           expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[c]' }] }`);
-         });
 
-
-      it('should delete the decorator (but cope with no trailing comma) that was matched in the analysis',
-         () => {
-           const {decorationAnalyses, sourceFile, renderer} = setup(PROGRAM);
-           const output = new MagicString(PROGRAM.contents);
-           const compiledClass =
-               decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'B') !;
-           const decorator = compiledClass.decorators[0];
-           const decoratorsToRemove = new Map<ts.Node, ts.Node[]>();
-           decoratorsToRemove.set(decorator.node.parent !, [decorator.node]);
-           renderer.removeDecorators(output, decoratorsToRemove);
-           expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[a]' }] },`);
-           expect(output.toString()).toContain(`{ type: OtherA }`);
-           expect(output.toString())
-               .not.toContain(`{ type: Directive, args: [{ selector: '[b]' }] }`);
-           expect(output.toString()).toContain(`{ type: OtherB }`);
-           expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[c]' }] }`);
-         });
-
-
-      it('should delete the decorator (and its container if there are no other decorators left) that was matched in the analysis',
-         () => {
-           const {decorationAnalyses, sourceFile, renderer} = setup(PROGRAM);
-           const output = new MagicString(PROGRAM.contents);
-           const compiledClass =
-               decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'C') !;
-           const decorator = compiledClass.decorators[0];
-           const decoratorsToRemove = new Map<ts.Node, ts.Node[]>();
-           decoratorsToRemove.set(decorator.node.parent !, [decorator.node]);
-           renderer.removeDecorators(output, decoratorsToRemove);
-           expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[a]' }] },`);
-           expect(output.toString()).toContain(`{ type: OtherA }`);
-           expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[b]' }] }`);
-           expect(output.toString()).toContain(`{ type: OtherB }`);
-           expect(output.toString())
-               .not.toContain(`{ type: Directive, args: [{ selector: '[c]' }] }`);
-           expect(output.toString()).not.toContain(`C.decorators = [`);
-         });
+    it('should delete the decorator (and following comma) that was matched in the analysis', () => {
+      const {renderer, decorationAnalyses, sourceFile} = setup(PROGRAM);
+      const output = new MagicString(PROGRAM.contents);
+      const compiledClass =
+          decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'A') !;
+      const decorator = compiledClass.decorators[0];
+      const decoratorsToRemove = new Map<ts.Node, ts.Node[]>();
+      decoratorsToRemove.set(decorator.node.parent !, [decorator.node]);
+      renderer.removeDecorators(output, decoratorsToRemove);
+      expect(output.toString()).not.toContain(`{ type: Directive, args: [{ selector: '[a]' }] },`);
+      expect(output.toString()).toContain(`{ type: OtherA }`);
+      expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[b]' }] }`);
+      expect(output.toString()).toContain(`{ type: OtherB }`);
+      expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[c]' }] }`);
     });
+
+
+    it('should delete the decorator (but cope with no trailing comma) that was matched in the analysis',
+       () => {
+         const {renderer, decorationAnalyses, sourceFile} = setup(PROGRAM);
+         const output = new MagicString(PROGRAM.contents);
+         const compiledClass =
+             decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'B') !;
+         const decorator = compiledClass.decorators[0];
+         const decoratorsToRemove = new Map<ts.Node, ts.Node[]>();
+         decoratorsToRemove.set(decorator.node.parent !, [decorator.node]);
+         renderer.removeDecorators(output, decoratorsToRemove);
+         expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[a]' }] },`);
+         expect(output.toString()).toContain(`{ type: OtherA }`);
+         expect(output.toString())
+             .not.toContain(`{ type: Directive, args: [{ selector: '[b]' }] }`);
+         expect(output.toString()).toContain(`{ type: OtherB }`);
+         expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[c]' }] }`);
+       });
+
+
+    it('should delete the decorator (and its container if there are not other decorators left) that was matched in the analysis',
+       () => {
+         const {renderer, decorationAnalyses, sourceFile} = setup(PROGRAM);
+         const output = new MagicString(PROGRAM.contents);
+         const compiledClass =
+             decorationAnalyses.get(sourceFile) !.compiledClasses.find(c => c.name === 'C') !;
+         const decorator = compiledClass.decorators[0];
+         const decoratorsToRemove = new Map<ts.Node, ts.Node[]>();
+         decoratorsToRemove.set(decorator.node.parent !, [decorator.node]);
+         renderer.removeDecorators(output, decoratorsToRemove);
+         renderer.addDefinitions(output, compiledClass, 'SOME DEFINITION TEXT');
+         expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[a]' }] },`);
+         expect(output.toString()).toContain(`{ type: OtherA }`);
+         expect(output.toString()).toContain(`{ type: Directive, args: [{ selector: '[b]' }] }`);
+         expect(output.toString()).toContain(`{ type: OtherB }`);
+         expect(output.toString()).toContain(`function C() {}\nSOME DEFINITION TEXT\n  return C;`);
+         expect(output.toString()).not.toContain(`C.decorators`);
+       });
+
   });
 
   describe('[__decorate declarations]', () => {
@@ -348,7 +405,7 @@ A.decorators = [
        });
 
 
-    it('should delete the decorator (and its container if there are not other decorators left) that was matched in the analysis',
+    it('should delete the decorator (and its container if there are no other decorators left) that was matched in the analysis',
        () => {
          const {renderer, decorationAnalyses, sourceFile} = setup(PROGRAM_DECORATE_HELPER);
          const output = new MagicString(PROGRAM_DECORATE_HELPER.contents);
@@ -364,7 +421,7 @@ A.decorators = [
          expect(output.toString()).toContain(`OtherB()`);
          expect(output.toString()).not.toContain(`Directive({ selector: '[c]' })`);
          expect(output.toString()).not.toContain(`C = tslib_1.__decorate([`);
-         expect(output.toString()).toContain(`let C = class C {\n};\nexport { C };`);
+         expect(output.toString()).toContain(`function C() {\n    }\n    return C;`);
        });
   });
 });
